refactor(Loader): client専用処理をclient/LoaderWrapperへ切り出しLoader自体から'use client'を除去する
ariaNotify呼び出しやMutationObserverなど、client専用APIに依存する
callbackRefのロジックはDOM要素にアタッチされるLoaderWrapper側の
関心事であり、Loader本体には不要だった。

このロジックをclient/LoaderWrapper.tsxへ移動することで、Loader自体は
memoとLocalizerのみを使うモジュールになり'use client'を除去できる。
実際にビルド出力でLoader.jsに'use client'が付かなくなり、
client/LoaderWrapper.jsにのみ付くことを確認済み。

hookが存在せずindex.tsからも公開されない内部コンポーネントのため、
client/直下にフラット配置する（CLAUDE.mdの判断基準）。
