+++ packages/smarthr-ui/src/components/Loader/Loader.tsx
@@ -1,12 +1,10 @@
-'use client'
-
-import { type ComponentProps, type ReactNode, memo, useCallback } from 'react'
+import { type ComponentProps, type ReactNode, memo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useCallbackRefCleanupForReact18 } from '../../hooks/client/useCallbackRefCleanupForReact18'
 import { Localizer } from '../../intl'
 
 import { LoaderSpinner } from './LoaderSpinner'
+import { LoaderWrapper } from './client'
 
 type BaseProps = {
   /** ローダーの大きさ */
@@ -22,7 +20,7 @@ type Props = BaseProps & Omit<ComponentProps<'span'>, keyof BaseProps>
 
 const classNameGenerator = tv({
   slots: {
-    wrapper: ['smarthr-ui-Loader', 'shr-inline-block shr-overflow-hidden'],
+    wrapper: ['smarthr-ui-Loader', 'shr-inline-block', 'shr-overflow-hidden'],
     textSlot: ['shr-block', 'shr-mt-1', 'shr-text-base', 'shr-text-center'],
   },
   variants: {
@@ -38,7 +36,7 @@ const classNameGenerator = tv({
 })
 
 export const Loader = memo<Props>(
-  ({ size = 'M', alt, text, type = 'primary', role = 'status', className, ...rest }) => {
+  ({ size = 'M', alt, text, type = 'primary', className, ...rest }) => {
     // HINT: Loaderは一度表示されれば属性が変わる可能性はほぼ無いためuseMemoしない
     const classNames = (() => {
       const { wrapper, textSlot } = classNameGenerator()
@@ -51,43 +49,15 @@ export const Loader = memo<Props>(
       }
     })()
 
-    const callbackRef = useCallbackRefCleanupForReact18(
-      useCallback(
-        (node: HTMLElement | null) => {
-          if (!node) {
-            return
-          }
-
-          const ariaNotifyAction = () => {
-            document.ariaNotify(node.innerText, { priority: role === 'alert' ? 'high' : 'normal' })
-          }
-
-          ariaNotifyAction()
-
-          const observer = new MutationObserver(ariaNotifyAction)
-          observer.observe(node, {
-            childList: true,
-            subtree: true,
-            characterData: true,
-          })
-
-          return () => {
-            observer.disconnect()
-          }
-        },
-        [role],
-      ),
-    )
-
     return (
-      <span {...rest} ref={callbackRef} className={classNames.wrapper}>
+      <LoaderWrapper {...rest} className={classNames.wrapper}>
         <LoaderSpinner
           type={type}
           alt={alt || <Localizer id="smarthr-ui/Loader/alt" defaultText="処理中" />}
           size={size}
         />
         {text && <span className={classNames.text}>{text}</span>}
-      </span>
+      </LoaderWrapper>
     )
   },
 )
