+++ packages/smarthr-ui/src/components/Loader/client/index.ts
@@ -1,0 +1,1 @@
+export { LoaderWrapper } from './LoaderWrapper'
